==> Lessons/10_DataVisualization.Rmd <==
---
title: "10: Data Visualization"
date: "Spring 2019" 14 Feb
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Perform advanced edits on ggplot objects to follow best practices for data visualization

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
PeterPaul.nutrients.gathered <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
EPAair <- read.csv("./Data/Processed/EPAair_O3PM25_3sites1718_processed.csv")

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

```


### Themes
Often, we will want to change multiple visual aspects of a plot. Ggplot comes with pre-built themes that will adjust components of plots if you call that theme. 
```{r, fig.height = 3, fig.width = 4, warning = FALSE}
O3plot <- ggplot(EPAair) + #ggpplot theme changes w/in 1 line, don;t have to redefine a bunch for each graph
  geom_point(aes(x = Date, y = Ozone)) 
print(O3plot) #default

O3plot1 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone)) +
  theme_gray() #add theme as if its a layer, this one is default
print(O3plot1)

O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone)) +
  theme_bw() #has box and white bkgrd
print(O3plot2)

O3plot3 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone)) +
  theme_light()
print(O3plot3) #all gray lines

O3plot4 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone)) +
  theme_classic()
print(O3plot4) #no grid, just black line for axes
```

Notice that some aspects of your graph have not been adjusted, including: 

* text size
* axis label colors
* legend position and justification

If you would like to set a common theme across all plots in your analysis session, you may define a theme and call up that theme for each graph. This eliminates the need to add multiple lines of code in each plot. 

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
mytheme <- theme_classic(base_size = 14) + #base font is 11 or 12
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top") #alternative: legend.position + legend.justification; can tell it to put legend in particular coord in graph 
#legend.position = c(1,1), legend.justification = c(1,1)

# options: call the theme in each plot or set the theme at the start. 

O3plot5 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone)) +
  mytheme #add as a layer
print(O3plot5)  

theme_set(mytheme) #if you're making a lot of plots for your whole session, can be overridden by specifiying w/in a single ggplot

O3plot6 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot6)  

```

### Adjusting multiple components of your plots

While the theme allows us to set multiple aspects of plots, ggplot allows us to adjust other parts of plots outside of the theme. 

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
O3plot7 <- ggplot(EPAair, aes(x = Date, y = Ozone)) +
  #geom_rect(xmin = as.Date("2016-01-01"), xmax = as.Date("2020-01-01"), #made background 0-50 green
            #ymin = 0, ymax = 50, fill = "green") +
  #geom_rect(xmin = as.Date("2016-01-01"), xmax = as.Date("2020-01-01"), 
            #ymin = 50, ymax = 100, fill = "yellow") + #makes 51-100 yellow
  geom_point() + #type of plot
  geom_text(x = as.Date("2019-01-01"), y = 45, label = "good", hjust = 1, fontface = "bold") + #labeled good at y=45 BUT you have to tell it to look for a date
  geom_text(x = as.Date("2019-01-01"), y = 95, label = "moderate", hjust = 1, fontface = "bold") +
  scale_x_date(limits = as.Date(c("2017-01-01", "2018-12-31")), #told it only to use month names as label instead of number
    date_breaks = "2 months", date_labels = "%b %y") + #broke up dates into every other month
  ylab(expression("O"[3]* " AQI Value")) + #brackets give you subscript for O3, superscript is a carat
  theme(axis.text.x = element_text(angle = 45,  hjust = 1)) #angles the text; hjust moves text horizontally; vjust for vertical
print(O3plot7)  

```

### Color palettes
There are several color palettes that are designed to be more effective than palettes in base R. These include Viridis (https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html) and Color Brewer (http://colorbrewer2.org/). A few rules for choosing colors: 

* Consider if your plot needs to be viewed in black and white. If so, choose a sequential palette with varying color intensity.
* Choose a palette that is color-blind friendly
* Maximize contrast (e.g., no pale colors on a white background)
* Diverging color palettes should be used for diverging values (e.g., warm-to-cool works well for values on a scale encompassing negative and positive values)

Perception is key! Choose palettes that are visually pleasing and will communicate what you are hoping your audience to perceive. Hint: base R palettes are not ideal.

```{r, warning = FALSE}
install.packages("viridis")
#install.packages("RColorBrewer")
install.packages("colormap")
library(viridis) #ff is last digits on hexcode
library(RColorBrewer) #diverging to make colorblind safe
library(colormap)

scales::show_col(colormap(colormap = colormaps$viridis, nshades = 16))
scales::show_col(colormap(colormap = colormaps$inferno, nshades = 16))
scales::show_col(colormap(colormap = colormaps$magma, nshades = 16))
display.brewer.all(n = 9)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth, shape = lakename)) +
  geom_point() 
print(NvsP)

# let's first make the plot look better.
# change your axis labels to reflect TN and TP in micrograms per liter.
# change your legend labels
NvsP2 <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, fill = depth, shape = lakename)) +
  geom_point(alpha = 0.8, size = 3,color = "magenta") +
xlab(expression("total phosphorus (\U003BCg/L)")) +
ylab(expression("total nitrogen (\U003BCg/L)")) +
#labs(x = "Total P", y = "Total N", fill = "Depth, m", shape = lakename)
  # change your legend labels here
  guides(fill=guide_legend(title="depth")) +
  scale_shape_manual(values = c(15, 17)) + #these are shape numbers; for other shapes you can define fill and outline
scale_color_distiller(palette = "Blues", direction = 1) + # use scale_color_brewer for discrete variables
#scale_color_viridis(option = "magma", direction = -1) +
  theme(legend.position = "left", 
        legend.text = element_text(size = 12), legend.title = element_text(size = 12))
print(NvsP2)

#from Kateri
NvsP2 <-
 ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, fill = depth, shape = lakename)) +
geom_point(alpha = 0.8, size = 3, color = “black”) +
xlab(“Total P (\U003BCg/L)“) +
ylab() +
 # change your legend labels here
 scale_shape_manual(values = c(21, 23)) +
 #scale_fill_distiller(palette = “Blues”, direction = 1) + # use scale_color_brewer for discrete variables
 scale_fill_viridis(option = “magma”, direction = -1) +
 theme(legend.position = “right”,
       legend.text = element_text(size = 12), legend.title = element_text(size = 12))

# change your y axis label to list concentration in micrograms per liter
# remove your x axis label
# change labels for nutrients in the legend
Nutrientplot <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration, color = nutrient)) +
  geom_boxplot() +
# place your additional edits here
  scale_y_continuous(expand = c(0, 0)) +
  #scale_color_brewer(palette = "YlGnBu") +
  #scale_color_manual(values = c("#7fcdbb", "#41b6c4", "#1d91c0", "#225ea8", "#0c2c84")) +
  #scale_color_viridis(discrete = TRUE) +
  theme(legend.position = "right")
print(Nutrientplot)
```

### Adjusting facets

```{r, fig.height = 4, fig.width = 6}
PMplot.faceted <-
  ggplot(EPAair, aes(x = month, y = PM2.5)) +
  geom_point() +
  facet_grid(Site.Name ~ year) + 
  scale_x_continuous(breaks = c(1:12)) +
  #theme(strip.background = element_rect(fill = "black"), strip.text = element_text(color = "white"))
  ylab(expression("PM 2.5 AQI Value")) 
print(PMplot.faceted)

```

### Multiple plots on a page
In situations where facets don't fill our needs to place multiple plots on a page, we can use the package `gridExtra` to arrange plots. The `grid.arrange` function is extremely flexible in its ability to arrange plots in specific configurations. A useful guide can be found here: https://cran.r-project.org/web/packages/egg/vignettes/Ecosystem.html.

```{r, fig.height = 7, fig.width = 6}
#install.packages("gridExtra")
library(gridExtra)
grid.arrange(NvsP2, Nutrientplot)
grid.arrange(O3plot7, PMplot.faceted)

```

### Saving plots
The `ggsave` function allows you to save plots in jpg, png, eps, pdf, tiff, and other formats. The following information can be supplied: 

* filename, with file extension and in quotes (required)
* plot object (required)
* path, with file name 
* width, height, units 
* resolution (dpi)

For example: 
`ggsave("PMplot.jpg", PMplot.faceted,  path = "./Output/PMplotjpg", height = 4, width = 6, units = "in", dpi = 300)`
